Code improvements and formatting Fixed wrong value in folder path handling Fixed incorrect filenames for Triforce files error messages Added socket value check in AMMBCommand::Closesocket Fixed incorrect array offsets in the network functions of the AMBaseboard Fixed not using the randomly generated filename when a backup file for Triforce is already in use Fixed DTV flag not being set when using NTSC-J

+++ Source/Core/Core/HW/SI/SI_DeviceAMBaseboard.cpp
@@ -245,7 +245,7 @@ void CSIDevice_AMBaseboard::ICCardSendReply(ICCommand* iccommand, u8* buffer, u3
 {
   iccommand->status = Common::swap16(iccommand->status);
 
-  const u16 crc = CheckSumXOR(iccommand->data + 2, iccommand->pktlen - 1);
+  const u8 crc = CheckSumXOR(iccommand->data + 2, iccommand->pktlen - 1);
 
   for (u32 i = 0; i < iccommand->pktlen + 1; ++i)
   {
